+++ tests/template/test_csrf_token.py
@@ -9,10 +9,10 @@ from starlite.config.csrf import CSRFConfig
 from starlite.config.template import TemplateConfig
 from starlite.contrib.jinja import JinjaTemplateEngine
 from starlite.contrib.mako import MakoTemplateEngine
+from starlite.middleware.csrf import generate_csrf_token
 from starlite.response_containers import Template
 from starlite.testing import create_test_client
 from starlite.types import Scope
-from starlite.utils import generate_csrf_token
 
 
 @pytest.mark.parametrize(
